Code exercise showing locations on google map that are stored in db.
Develop an HTML5-based web application that plots an arbitrary set of
geographic points on a map using a common mapping library such as
Google Maps or OpenStreet Maps. The points should be clickable so that
further HTML content can be displayed about each point, and the map
should support common activities such as panning and zooming. The
geographic points and accompanied popup HTML content must be stored in
a backend database of your choice such as MySQL, Sqlite or MongoDB, and
a PHP script must be used to read the data and produce AJAX-loadable
content for the front-end Javascript application to interpret.

You may use Web research to inform your work, but all non-API code must
be your own. When you are complete, please share with us a URL to the
working application and commit all code to a GitHub project and share
that URL with us as well.